+++ integrationtests/Paket.IntegrationTests/BasicResolverSpecs.fs
@@ -113,9 +113,10 @@ let ``#1189 should allow # in path``() =
     lockFile.Groups.[Constants.MainDependencyGroup].Resolution.[PackageName "FAKE"].Version
     |> shouldBeGreaterThan (SemVer.Parse "4.7.2")
 
+
 [<Test>]
 let ``#1247 shouldn't load lockfile in full update``() =
-    update "i001247-lockfile-error" |> ignore
+    updateEx true "i001247-lockfile-error" |> ignore
 
 [<Test>]
 let ``#1247 should report lockfile in parse errror``() =
@@ -137,7 +138,10 @@ let ``#1450 should resolve with twiddle wakka``() =
     let lockFile = update "i001450-twiddle-wakka"
     lockFile.Groups.[Constants.MainDependencyGroup].Resolution.[PackageName "EnterpriseLibrary.SemanticLogging"].Version
     |> shouldBeSmallerThan (SemVer.Parse "3")
-
+    
+[<Test>]
+let ``#2640 shouldn't try GetDetails if package only exists locally``() =
+    updateEx true "i002640" |> ignore
 
 #if INTERACTIVE
 ;;