+++ integrationtests/Paket.IntegrationTests/TestHelper.fs
@@ -35,6 +35,10 @@ let partitionForTravis scenario =
     
 
 let paketToolPath = FullName(__SOURCE_DIRECTORY__ + "../../../bin/paket.exe")
+let dotnetToolPath =
+    match Environment.GetEnvironmentVariable "DOTNET_EXE_PATH" with
+    | null | "" -> "dotnet"
+    | s -> s
 let integrationTestPath = FullName(__SOURCE_DIRECTORY__ + "../../../integrationtests/scenarios")
 let scenarioTempPath scenario = Path.Combine(integrationTestPath,scenario,"temp")
 let originalScenarioPath scenario = Path.Combine(integrationTestPath,scenario,"before")
@@ -70,38 +74,58 @@ let prepare scenario =
         for file in Directory.GetFiles(scenarioPath, (sprintf "*.%stemplate" ext), SearchOption.AllDirectories) do
             File.Move(file, Path.ChangeExtension(file, ext))
 
-let directPaketInPath command scenarioPath =
+let prepareSdk scenario =
+    let tmpPaketFolder = (scenarioTempPath scenario) @@ ".paket"
+    let targetsFile = FullName(__SOURCE_DIRECTORY__ + "../../../src/Paket/embedded/Paket.Restore.targets")
+    let paketExe = FullName(__SOURCE_DIRECTORY__ + "../../../bin/paket.exe")
+
+    setEnvironVar "PaketExePath" paketExe
+    prepare scenario
+    if (not (Directory.Exists tmpPaketFolder)) then
+        Directory.CreateDirectory tmpPaketFolder |> ignore
+
+    FileHelper.CopyFile tmpPaketFolder targetsFile
+
+
+type PaketMsg =
+  { IsError : bool; Message : string }
+    static member isError ({ IsError = e}:PaketMsg) = e
+    static member getMessage ({ Message = msg }:PaketMsg) = msg
+
+let directToolEx isPaket toolPath command workingDir =
     #if INTERACTIVE
     let result =
         ExecProcessWithLambdas (fun info ->
-          info.FileName <- paketToolPath
-          info.WorkingDirectory <- scenarioPath
+          info.FileName <- toolPath
+          info.WorkingDirectory <- workingDir
           info.Arguments <- command) 
           (System.TimeSpan.FromMinutes 7.)
           false
           (printfn "%s")
           (printfn "%s")
-    string result
+    let res = new ResizeArray()
+    res.Add (string result)
+    res
     #else
     Environment.SetEnvironmentVariable("PAKET_DETAILED_ERRORS", "true")
-    printfn "%s> paket %s" scenarioPath command
+    printfn "%s> %s %s" workingDir (if isPaket then "paket" else toolPath) command
     let perfMessages = ResizeArray()
-    let msgs = ResizeArray()
+    let msgs = ResizeArray<PaketMsg>()
     let mutable perfMessagesStarted = false
     let addAndPrint isError msg =
         if not isError then
-            if msg = "Performance:" then
+            if isPaket && msg = "Performance:" then
                 perfMessagesStarted <- true
-            elif perfMessagesStarted then
+            elif isPaket && perfMessagesStarted then
                 perfMessages.Add(msg)
-                
-        msgs.Add((isError, msg))
+
+        msgs.Add({ IsError = isError; Message = msg})
         
     let result =
         try
             ExecProcessWithLambdas (fun info ->
-              info.FileName <- paketToolPath
-              info.WorkingDirectory <- scenarioPath
+              info.FileName <- toolPath
+              info.WorkingDirectory <- workingDir
               info.CreateNoWindow <- true
               info.Arguments <- command)
               (System.TimeSpan.FromMinutes 7.)
@@ -114,61 +138,92 @@ let directPaketInPath command scenarioPath =
             else
                 printfn "ExecProcessWithLambdas failed. Output was: "
 
-            for isError, msg in msgs do
+            for { IsError = isError; Message = msg } in msgs do
                 printfn "%s%s" (if isError then "ERR: " else "") msg
             reraise()
-    // Only throw after the result <> 0 check because the current test might check the argument parsing
-    // this is the only case where no performance is printed
-    let isUsageError = result <> 0 && msgs |> Seq.filter fst |> Seq.map snd |> Seq.exists (fun msg -> msg.Contains "USAGE:")
-    if not isUsageError then
-        if perfMessages.Count = 0 then
-            failwith "No Performance messages recieved in test!"
-        printfn "Performance:"
-        for msg in perfMessages do
-            printfn "%s" msg
+    if isPaket then
+        // Only throw after the result <> 0 check because the current test might check the argument parsing
+        // this is the only case where no performance is printed
+        let isUsageError = result <> 0 && msgs |> Seq.filter PaketMsg.isError |> Seq.map PaketMsg.getMessage |> Seq.exists (fun msg -> msg.Contains "USAGE:")
+        if not isUsageError then
+            if perfMessages.Count = 0 then
+                failwith "No Performance messages recieved in test!"
+            printfn "Performance:"
+            for msg in perfMessages do
+                printfn "%s" msg
 
     // always print stderr
-    for isError, msg in msgs do
-        if isError then
-            printfn "ERR: %s" msg
+    for msg in msgs do
+        if msg.IsError then
+            printfn "ERR: %s" msg.Message
 
     if result <> 0 then 
-        let errors = String.Join(Environment.NewLine,msgs |> Seq.filter fst |> Seq.map snd)
+        let errors = String.Join(Environment.NewLine,msgs |> Seq.filter PaketMsg.isError |> Seq.map PaketMsg.getMessage)
         if String.IsNullOrWhiteSpace errors then
             failwithf "The process exited with code %i" result
         else
             failwith errors
 
-
-    String.Join(Environment.NewLine,msgs |> Seq.map snd)
+    msgs
     #endif
 
-let directPaket command scenario =
-    partitionForTravis scenario
-    directPaketInPath command (scenarioTempPath scenario)
+let directPaketInPathEx command scenarioPath =
+    directToolEx true paketToolPath command scenarioPath
 
-let paket command scenario =
+let checkResults msgs =
+    msgs
+    |> Seq.filter PaketMsg.isError
+    |> Seq.toList
+    |> shouldEqual []
+
+let directDotnet checkZeroWarn command workingDir =
+    let msgs = directToolEx false dotnetToolPath command workingDir
+    if checkZeroWarn then checkResults msgs
+    msgs
+
+let private fromMessages msgs =
+    String.Join(Environment.NewLine,msgs |> Seq.map PaketMsg.getMessage)
+
+let directPaketInPath command scenarioPath = directPaketInPathEx command scenarioPath |> fromMessages
+
+let directPaketEx command scenario =
+    partitionForTravis scenario
+    directPaketInPathEx command (scenarioTempPath scenario)
+
+let directPaket command scenario = directPaketEx command scenario |> fromMessages
+
+let paketEx checkZeroWarn command scenario =
     prepare scenario
 
-    directPaket command scenario
+    let msgs = directPaketEx command scenario
+    if checkZeroWarn then checkResults msgs
+    msgs
 
-let update scenario =
+let paket command scenario =
+    paketEx false command scenario |> fromMessages
+
+let updateEx checkZeroWarn scenario =
     #if INTERACTIVE
     paket "update --verbose" scenario |> printfn "%s"
     #else
-    paket "update" scenario |> ignore
+    paketEx checkZeroWarn "update" scenario |> ignore
     #endif
     LockFile.LoadFrom(Path.Combine(scenarioTempPath scenario,"paket.lock"))
 
-let install scenario =
+let update scenario =
+    updateEx false scenario
+
+let installEx checkZeroWarn scenario =
     #if INTERACTIVE
     paket "install --verbose" scenario |> printfn "%s"
     #else
-    paket "install" scenario |> ignore
+    paketEx checkZeroWarn  "install" scenario |> ignore
     #endif
     LockFile.LoadFrom(Path.Combine(scenarioTempPath scenario,"paket.lock"))
 
-let restore scenario = paket "restore" scenario |> ignore
+let install scenario = installEx false scenario
+
+let restore scenario = paketEx false "restore" scenario |> ignore
 
 let updateShouldFindPackageConflict packageName scenario =
     try
@@ -180,3 +235,39 @@ let updateShouldFindPackageConflict packageName scenario =
         printfn "Ninject conflict test passed"
         #endif
         ()
+
+let clearPackage name =
+    // ~/.nuget/packages
+    let userPackageFolder = Paket.Constants.UserNuGetPackagesFolder
+
+    // %APPDATA%/NuGet/Cache
+    let nugetCache = Paket.Constants.NuGetCacheFolder
+
+    for cacheDir in [ nugetCache; userPackageFolder ] do
+        if Directory.Exists cacheDir then
+            Directory.EnumerateDirectories(cacheDir)
+            |> Seq.filter (fun n -> Path.GetFileName n |> String.startsWithIgnoreCase name)
+            |> Seq.iter (fun n -> Directory.Delete(n, true))
+            Directory.EnumerateFiles(cacheDir)
+            |> Seq.filter (fun n -> Path.GetFileName n |> String.startsWithIgnoreCase name)
+            |> Seq.iter (fun n -> File.Delete(n))
+
+let isPackageCached name =
+    // ~/.nuget/packages
+    let userPackageFolder = Paket.Constants.UserNuGetPackagesFolder
+
+    // %APPDATA%/NuGet/Cache
+    let nugetCache = Paket.Constants.NuGetCacheFolder
+
+    [ nugetCache; userPackageFolder ]
+    |> List.collect (fun cacheDir ->
+        if Directory.Exists cacheDir then
+            let dirs =
+                Directory.EnumerateDirectories(cacheDir)
+                |> Seq.filter (fun n -> Path.GetFileName n |> String.startsWithIgnoreCase name)
+            let files =
+                Directory.EnumerateFiles(cacheDir)
+                |> Seq.filter (fun n -> Path.GetFileName n |> String.startsWithIgnoreCase name)
+            Seq.append dirs files
+            |> Seq.toList
+        else [])