+++ src/Paket.Core/Versioning/PackageSources.fs
@@ -81,81 +81,8 @@ type NugetSource =
     member x.BasicAuth =
         x.Authentication |> Option.map toCredentials
 
-type NugetV3SourceResourceJSON = 
-    { [<JsonProperty("@type")>]
-      Type : string
-      [<JsonProperty("@id")>]
-      ID : string }
-
-type NugetV3SourceRootJSON = 
-    { [<JsonProperty("resources")>]
-      Resources : NugetV3SourceResourceJSON [] }
-
 type NugetV3Source = NugetSource
-//type NugetV3Source = 
-//    { Url : string
-//      Authentication : NugetSourceAuthentication option }
 
-type NugetV3ResourceType = 
-    | AutoComplete
-    | AllVersionsAPI
-    | Registration
-
-    member this.AsString = 
-        match this with
-        | AutoComplete -> "SearchAutoCompleteService"
-        | Registration -> "RegistrationsBaseUrl"
-        | AllVersionsAPI -> "PackageBaseAddress/3.0.0"
-
-// Cache for nuget indices of sources
-type ResourceIndex = Map<NugetV3ResourceType,string>
-let private nugetV3Resources = System.Collections.Concurrent.ConcurrentDictionary<NugetV3Source,Task<ResourceIndex>>()
-
-let getNuGetV3Resource (source : NugetV3Source) (resourceType : NugetV3ResourceType) : Async<string> =
-    let key = source
-    let getResourcesRaw () =
-        async {
-            let basicAuth = source.Authentication |> Option.map toCredentials
-            let! rawData = safeGetFromUrl(basicAuth, source.Url, acceptJson)
-            let rawData =
-                match rawData with
-                | NotFound ->
-                    raise <| new Exception(sprintf "Could not load resources (404) from '%s'" source.Url)
-                | UnknownError e ->
-                    raise <| new Exception(sprintf "Could not load resources from '%s'" source.Url, e.SourceException)
-                | SuccessResponse x -> x
-
-            let json = JsonConvert.DeserializeObject<NugetV3SourceRootJSON>(rawData)
-            let resources =
-                json.Resources
-                |> Seq.distinctBy(fun x -> x.Type.ToLower())
-                |> Seq.map(fun x -> x.Type.ToLower(), x.ID)
-            let map =
-                resources
-                |> Seq.choose (fun (res, value) ->
-                    let resType =
-                        match res.ToLower() with
-                        | "searchautocompleteservice" -> Some AutoComplete
-                        | "registrationsbaseurl" -> Some Registration
-                        | "packagebaseaddress/3.0.0" -> Some AllVersionsAPI
-                        | _ -> None
-                    match resType with
-                    | None -> None
-                    | Some k ->
-                        Some (k, value))
-                |> Seq.distinctBy fst
-                |> Map.ofSeq
-            return map
-        } |> Async.StartAsTask
-
-    async {
-        let t = nugetV3Resources.GetOrAdd(key, (fun _ -> getResourcesRaw()))
-        let! res = t |> Async.AwaitTask
-        return
-            match res.TryFind resourceType with
-            | Some s -> s
-            | None -> failwithf "could not find an %s endpoint for %s" (resourceType.ToString()) source.Url
-    }
 let userNameRegex = Regex("username[:][ ]*[\"]([^\"]*)[\"]", RegexOptions.IgnoreCase ||| RegexOptions.Compiled)
 let passwordRegex = Regex("password[:][ ]*[\"]([^\"]*)[\"]", RegexOptions.IgnoreCase ||| RegexOptions.Compiled)
 let authTypeRegex = Regex("authtype[:][ ]*[\"]([^\"]*)[\"]", RegexOptions.IgnoreCase ||| RegexOptions.Compiled)